ambient-light/AmbientLightSensor_start.https.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>AmbientLightSensor Test: start()</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="help" href="https://www.w3.org/TR/ambient-light/">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>

let sensor, start_return;

setup(function() {
  sensor = new AmbientLightSensor();
});

test(function() {
  assert_equals(sensor.reading, null);
}, "The default sensor.reading is 'null'");

test(function() {
  assert_equals(sensor.state, "idle");
}, "The default sensor.state is 'idle'");

async_test(function(t) {
  sensor.onstatechange = t.step_func_done(function(event) {
    assert_equals(sensor.state, "activating");
  });
  sensor.onerror = t.step_func_done(function(event) {
    assert_unreached(event.error.name + ":" + event.error.message);
  });
  start_return = sensor.start();
}, "The sensor.state changes to 'activating' after sensor.start()");

test(function() {
  assert_equals(String(sensor.reading), "[object AmbientLightSensorReading]");
}, "the sensor.reading is AmbientLightSensorReading after executing start() method");

test(function() {
  assert_throws("InvalidStateError", function() { sensor.start(); }, "start() twice");
}, "throw an InvalidStateError exception when state is neither idle nor errored");

//TODO: The permission is not ready.

test(function() {
  assert_equals(start_return, undefined);
}, "the sensor.start() return undefined");

</script>
